Revert FDSC to 0.14.1 code until MS auth is fixed
1.2.1 has an issue with MS API authentication.  0.14.1 is the last confirmed usable revision.

This PR does a version bump to 1.2.2, but reverts to 0.14.1's code for the most part.  This will permit updates, when forced by Tampermonkey or Greasemonkey whether the URLs for download/update are removed or not from the script, to forcibly downgrade to 0.14.1's working code.

Any other changes after the fact to fix issues should revert this commit first.

Note that I don't propose this lightly - I hate suggesting a full revert to older code, but newer users are getting nonworking code, and existing users who can't disable script updates (Tampermonkey on Chrome for example) are getting forced to code that doesn't work - this is more or less a 'last resort' move to force the use of the older code until newer fixed code is available.