FIRE: Use CSS in injectMainCSS() within fire.user.js instead of SCSS
All of developing, testing, and releasing the userscript are substantially
easier without having the external dependency of needing both the SCSS->CSS
translation and making sure the hosed version of the CSS matches the version
of the code which is currently being used (i.e. generally requiring separate
hosting for the production CSS and any test CSS).

In the future, it's fine to write things in SCSS, but the controlling code is
in the fire.user.js file.