SIM: Make "smokey" quicklink a proper link.
This changeset converts the "smokey" quicklink that is added
to posts from a `<button>` to an `<a>` element in order to
improve consistency (the expected behavior of things that
look like links is for them to function like links).

It retains the exact same styling. In the native SE UI,
"share" and "edit" are `<a>` elements, whereas "close",
"delete", and "flag" are `<button>` elements. The `<a>`
elements show a pop-up window on a normal click interaction.
However, they can also be clicked with a modifier key (and/or
interacted with using the context menu) to make them work
like a normal hyperlink.

So, the "smokey" quicklink continues to work as it always has,
showing a pop-up when clicked normally, but it now also works
as a proper hyperlink in certain contexts. The hyperlink goes
directly to the MetaSmoke page for the post. Obviously, it
will not load anything if the post was not reported to MS.